Rename in redux-global-selectors
+ rename react-redux-with-selectors in react-redux-global-selectors
