+++ CONTRIBUTING.md
@@ -1,9 +1,9 @@
 # Contributing
 
-[react-redux-with-selectors](https://github.com/guillaumearm/react-redux-with-selectors) use [redux-with-selectors](https://github.com/guillaumearm/redux-with-selectors) as main dependency, so we work on a [dev branch](https://github.com/guillaumearm/redux-with-selectors/tree/dev) for staging and need to follow this __3 principles__ :
+[react-redux-global-selectors](https://github.com/guillaumearm/react-redux-global-selectors) use [redux-global-selectors](https://github.com/guillaumearm/redux-global-selectors) as main dependency, so we work on a [dev branch](https://github.com/guillaumearm/redux-global-selectors/tree/dev) for staging and need to follow this __3 principles__ :
 - this branch must be as clean as master, so `npm run prepublish` must be ok.
 - all master merges should be tagged by a new version. (on the master branch only)
-- all change must be validated by [react-redux-with-selectors](https://github.com/guillaumearm/react-redux-with-selectors) unit testing before merge on master
+- all change must be validated by [react-redux-global-selectors](https://github.com/guillaumearm/react-redux-global-selectors) unit testing before merge on master
 
 --------------
 
@@ -25,7 +25,7 @@
 
         $ git push origin <branch>
 
-6.  Open a pull request on [dev](https://github.com/guillaumearm/redux-with-selectors/tree/dev) branch.
+6.  Open a pull request on [dev](https://github.com/guillaumearm/redux-global-selectors/tree/dev) branch.
 
 ----------------
