+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "redux-with-selectors",
+  "name": "redux-global-selectors",
   "version": "0.4.0",
   "description": "a redux store enhancer for adding selectors in the store.",
-  "main": "./dist/redux-with-selectors.js",
+  "main": "./dist/redux-global-selectors.js",
   "scripts": {
     "build": "babel src --out-dir dist --ignore '*.spec.js'",
     "clean": "git reset HEAD -- ./dist ; git checkout ./dist",
@@ -18,7 +18,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/guillaumearm/redux-with-selectors.git"
+    "url": "https://github.com/guillaumearm/redux-global-selectors.git"
   },
   "keywords": [
     "redux",
@@ -28,9 +28,9 @@
   "author": "Guillaume ARM <garm@student.42.fr>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/guillaumearm/redux-with-selectors/issues"
+    "url": "https://github.com/guillaumearm/redux-global-selectors/issues"
   },
-  "homepage": "https://github.com/guillaumearm/redux-with-selectors",
+  "homepage": "https://github.com/guillaumearm/redux-global-selectors",
   "jest": {
     "moduleFileExtensions": [
       "js"
