+++ src/redux-global-selectors.js
