+++ src/withSelectors.js
@@ -1,7 +1,7 @@
 import { isObject, isFunction, flattenObject } from './internal';
 
-const MUST_BE_AN_OBJECT = 'redux-with-selectors : selectors must be a plain object.';
-const MUST_BE_A_FUNC = 'redux-with-selectors : a selector must be a function.';
+const MUST_BE_AN_OBJECT = 'redux-global-selectors : selectors must be a plain object.';
+const MUST_BE_A_FUNC = 'redux-global-selectors : a selector must be a function.';
 const error = msg => { throw new Error(msg) };
 
 const checkSelectors = s => {
